requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.11

# Configuration
django-environ==0.4.0



# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.6

# Images
Pillow==6.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.5.0
redis>=2.10.5


# Your custom requirements go here
python-docx==0.8.6
django-taggit==0.24
django-haystack==2.8.0
djangorestframework==3.7.0
whoosh==2.7.4
djangorestframework-jwt==1.11.0
drf-haystack==1.8.1
django-rest-swagger==2.1.1
markdown==2.6.7
django-rest-auth==0.9.5
django-cors-headers==2.0.1
elasticsearch>=2.0,<3.0
djangorestframework-recursive==0.1.2
django-import-export==1.2.0
celery==4.2.1
beautifulsoup4==4.7.1
fuzzywuzzy==0.17.0
django-silk==3.0.2
